:sparkles: phrase型のbookIDの型をintegerに変更

+++ SparkleClient/Classes/OpenAPIs/Models/Phrase.swift
@@ -13,11 +13,11 @@ public struct Phrase: Codable {
 
     public var createdAt: String?
     public var phrase: String?
-    public var bookId: Double
+    public var bookId: Int
     public var userID: UserInPhrases?
     public var updatedAt: String?
 
-    public init(createdAt: String?, phrase: String?, bookId: Double, userID: UserInPhrases?, updatedAt: String?) {
+    public init(createdAt: String?, phrase: String?, bookId: Int, userID: UserInPhrases?, updatedAt: String?) {
         self.createdAt = createdAt
         self.phrase = phrase
         self.bookId = bookId
